:construction: support detecting shell on windows
