+++ .vscode/settings.json
@@ -1,0 +1,5 @@
+{
+  "workbench.colorCustomizations": {
+    "terminal.background": "#000070"
+  },
+}
